+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./nack-codes.js";
+export * from "./nack-errors/nack-codes.js";
 export * from "./BulkInputData.js";
 export * from "./DebugGroup.js";
 export * from "./DioDirection.js";
@@ -15,3 +15,16 @@ export * from "./SerialFlowControl.js";
 export * from "./SerialParity.js";
 export * from "./VerbosityLevel.js";
 export * from "./Version.js";
+export * from "./nack-errors/NackError.js";
+export * from "./nack-errors/BatteryTooLowError.js";
+export * from "./nack-errors/diagnostic-errors.js";
+export * from "./nack-errors/digital-channel-errors.js";
+export * from "./nack-errors/GeneralSerialError.js";
+export * from "./nack-errors/i2c-errors.js";
+export * from "./nack-errors/motor-errors.js";
+export * from "./nack-errors/nack-code-to-error.js";
+export * from "./nack-errors/NoExpansionHubWithAddressError.js";
+export * from "./nack-errors/ParameterOutOfRangeError.js";
+export * from "./nack-errors/servo-errors.js";
+export * from "./nack-errors/TimeoutError.js";
+export * from "./nack-errors/UnrecognizedNackError.js";
