+++ packages/core/src/nack-errors/BatteryTooLowError.ts
