+++ packages/core/src/nack-errors/GeneralSerialError.ts
