+++ packages/core/src/nack-errors/i2c-errors.ts
@@ -1,5 +1,5 @@
 import { NackError, setPrototypeOf } from "./NackError.js";
-import { NackCode } from "@rev-robotics/rev-hub-core";
+import { NackCode } from "./nack-codes.js";
 
 export class I2cControllerBusyError extends NackError {
     constructor() {
