+++ packages/core/src/nack-errors/servo-errors.ts
@@ -1,6 +1,6 @@
 import { NackError, setPrototypeOf } from "./NackError.js";
 import { BatteryTooLowError } from "./BatteryTooLowError.js";
-import { NackCode } from "@rev-robotics/rev-hub-core";
+import { NackCode } from "./nack-codes.js";
 
 export class ServoNotFullyConfiguredError extends NackError {
     constructor() {
