+++ packages/expansion-hub/src/errors/RhspLibError.ts
@@ -1,4 +1,4 @@
-import { setPrototypeOf } from "./NackError.js";
+import { setPrototypeOf } from "@rev-robotics/rev-hub-core";
 
 export class RhspLibError extends Error {
     constructor(message: string) {
