+++ packages/expansion-hub/src/index.ts
@@ -1,18 +1,5 @@
-export { NackError } from "./errors/NackError.js";
-export * from "./errors/UnrecognizedNackError.js";
-export * from "./errors/i2c-errors.js";
-export * from "./errors/ParameterOutOfRangeError.js";
-export * from "./errors/diagnostic-errors.js";
-export * from "./errors/BatteryTooLowError.js";
-export * from "./errors/digital-channel-errors.js";
-export * from "./errors/motor-errors.js";
-export * from "./errors/servo-errors.js";
-export * from "./errors/nack-code-to-error.js";
-export * from "./errors/NoExpansionHubWithAddressError.js";
-export * from "./errors/RevHubError.js";
+export { NackError } from "@rev-robotics/rev-hub-core";
 export * from "./errors/RhspLibError.js";
-export * from "./errors/GeneralSerialError.js";
-export * from "./errors/TimeoutError.js";
 export * from "./discovery.js";
 export * from "./open-rev-hub.js";
 export * from "./led-pattern.js";
